requirements.txt
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
configparser==5.2.0
feedparser==6.0.10
future==0.18.2
idna==2.10
multidict==5.0.0
oauthlib==3.1.0
python-dotenv==0.19.2
python-twitter==3.5
requests==2.24.0
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
